debugging só sleepy

+++ backup/backup.sh
@@ -7,29 +7,13 @@ function backup() {
         mkdirprint "$2";
     fi
     for file in "$1"/*; do
-        if [[ -d "$file" ]]; then
-            backup "$file" "$2/$(basename "$file")"
-            continue;
-        elif [[ ! "$(basename "$file")" =~ $REGEX ]]; then
-            echo "$(basename $2) doesnt match regex"
-            continue;
-        fi
-        cpprint "$file" "$2/$(basename "$file")"
-    done
-}
-
-function backup_some() {
-    if [[ ! -d "$2" ]]; then
-        mkdirprint "$2";
-    fi
-    for file in "$1"/*; do
         if is_in_list "$file" "${DIRS[@]}" ; then
             continue;
         fi
         if [[ -d "$file" ]]; then
             backup "$file" "$2/$(basename "$file")"
+            continue;
         elif [[ ! "$(basename "$file")" =~ $REGEX ]]; then
-            echo "$(basename $2) doesnt match regex"
             continue;
         fi
         cpprint "$file" "$2/$(basename "$file")"
@@ -41,6 +25,9 @@ function backup_delete() {
         return 0;
     fi
     for file in "$2"/*; do
+        if is_in_list "$file" "${DIRS[@]}" ; then
+            continue;
+        fi
         if [[ -d "$file" ]]; then
             if [[ ! -d "$1/$(basename "$file")" ]]; then
                 rm -rf "$file"
@@ -49,27 +36,6 @@ function backup_delete() {
             backup_delete "$1/$(basename "$file")" "$2/$(basename "$file")"
             continue;
         fi
-        if [[ ! -f "$file" || -f "$1/$(basename "$file")" ]]; then
-            continue;
-        fi
-        if [[ $CHECKING -eq "0" ]]; then
-            rm "$file"
-        fi
-    done
-}
-
-function backup_delete_some() {
-    if [[ ! -d "$2" || ! -n "$2" ]]; then
-        return 0;
-    fi
-    for file in "$2"/*; do
-        if is_in_list "$file" "${DIRS[@]}" ; then
-            continue;
-        fi
-        if [[ -d "$file" ]]; then
-            backup_delete "$file" "$2/$(basename "$file")"
-            continue;
-        fi 
         if [[ ! -f "$file" || -f "$1/$(basename "$file")" ]]; then
             continue;
         fi
@@ -109,6 +75,9 @@ while getopts "cb:r:" opt; do
         r)
             REGEX="$OPTARG"
             check_regex "$REGEX"
+            if [[ $? -eq 1 ]]; then
+                exit 1
+            fi
             ;;
         \?)
             echo "Invalid option: -$OPTARG"
@@ -145,10 +114,5 @@ while [[ "$BACKUP_PATH" != "/" ]]; do
     BACKUP_PATH="$(dirname "$BACKUP_PATH")"
 done
 
-if [[ -z $DIRS_FILE ]]; then
-    backup_delete "$WORKDIR" "$BACKUP"
-    backup "$WORKDIR" "$BACKUP"
-else
-    backup_delete_some "$WORKDIR" "$BACKUP"
-    backup_some "$WORKDIR" "$BACKUP"
-fi
+backup_delete "$WORKDIR" "$BACKUP"
+backup "$WORKDIR" "$BACKUP"
